# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hashr",
    "version": "0.3.0",
    "description": "Machine-payable compute server that supports AI agents using the x402 payment protocol."
  },
  "paths": {
    "/compute/hash": {
      "post": {
        "summary": "Hash input payload",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"type": "object"}
            }
          }
        },
        "responses": {
          "200": {"description": "Hash computed"},
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/PaymentRequired"}
              }
            }
          }
        }
      }
    },
    "/compute/search": {
      "post": {
        "summary": "Search query",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"type": "object"}
            }
          }
        },
        "responses": {
          "200": {"description": "Search executed"},
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/PaymentRequired"}
              }
            }
          }
        }
      }
    },
    "/compute/inference": {
      "post": {
        "summary": "Inference request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"type": "object"}
            }
          }
        },
        "responses": {
          "200": {"description": "Inference executed"},
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/PaymentRequired"}
              }
            }
          }
        }
      }
    },
    "/job": {
      "post": {
        "summary": "Create a job",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"type": "object"}
            }
          }
        },
        "responses": {"200": {"description": "Job created"}}
      }
    },
    "/job/{id}": {
      "get": {
        "summary": "Get job",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Job retrieved"}}
      }
    },
    "/job/{id}/execute": {
      "post": {
        "summary": "Execute job",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"description": "Job executed"},
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/PaymentRequired"}
              }
            }
          }
        }
      }
    },
    "/agent/register": {
      "post": {
        "summary": "Register agent",
        "requestBody": {"required": true,"content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {"200": {"description": "Agent registered"}}
      }
    },
    "/reputation/query": {
      "post": {
        "summary": "Query reputation",
        "requestBody": {"required": true,"content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {"200": {"description": "Reputation returned"}}
      }
    },
    "/agent/{id}": {
      "get": {
        "summary": "Get agent",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Agent retrieved"}}
      }
    },
    "/reputation/erc8004": {
      "post": {
        "summary": "ERC8004 reputation profile",
        "requestBody": {"required": true,"content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {"200": {"description": "Profile returned"}}
      }
    },
    "/proof/verify": {
      "post": {
        "summary": "Verify proof",
        "requestBody": {"required": true,"content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {
          "200": {"description": "Proof verified"},
          "402": {
            "description": "Payment Required",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaymentRequired"}}}
          }
        }
      }
    },
    "/proof/{job_id}": {
      "get": {
        "summary": "Get proof",
        "parameters": [{"name": "job_id","in":"path","required":true,"schema":{"type":"string"}}],
        "responses": {
          "200": {"description": "Proof returned"},
          "402": {
            "description": "Payment Required",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaymentRequired"}}}
          }
        }
      }
    },
    "/wallet/risk": {
      "post": {
        "summary": "Wallet risk scan",
        "requestBody": {"required": true,"content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {
          "200": {"description": "Risk returned"},
          "402": {"description": "Payment Required","content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaymentRequired"}}}}
        }
      }
    },
    "/agent/identity": {
      "post": {
        "summary": "Register agent identity",
        "requestBody": {"required": true,"content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {"200": {"description": "Identity registered"}}
      }
    },
    "/agent/sign": {
      "post": {
        "summary": "Verify signature",
        "requestBody": {"required": true,"content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {
          "200": {"description": "Signature verified"},
          "402": {"description": "Payment Required","content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaymentRequired"}}}}
        }
      }
    },
    "/agent/{id}/public-key": {
      "get": {
        "summary": "Get agent public key",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Public key returned"}}
      }
    }
  },
  "components": {
    "schemas": {
      "PaymentRequired": {
        "type": "object",
        "properties": {
          "price": {"type": "number"},
          "payment_address": {"type": "string"},
          "job_id": {"type": "string"}
        }
      }
    }
  }
}